[keyboard] align with file and event naming conventions
