+++ packages/core/src/browser/keyboard/keyboard-layout-service.spec.ts
@@ -18,7 +18,7 @@ import { Container, injectable } from 'inversify';
 import { Emitter } from '../../common/event';
 import { KeyCode } from './keys';
 import { KeyboardLayoutService } from './keyboard-layout-service';
-import { KeyboardLayoutProvider, NativeKeyboardLayout, KeyboardLayoutChangeNotifier } from '../../common/keyboard/layout-provider';
+import { KeyboardLayoutProvider, NativeKeyboardLayout, KeyboardLayoutChangeNotifier } from '../../common/keyboard/keyboard-layout-provider';
 import * as os from '../../common/os';
 import * as chai from 'chai';
 import * as sinon from 'sinon';
@@ -47,7 +47,7 @@ describe('keyboard layout service', function () {
         @injectable()
         class MockLayoutProvider implements KeyboardLayoutProvider, KeyboardLayoutChangeNotifier {
             emitter = new Emitter<NativeKeyboardLayout>();
-            get onNativeLayoutChanged() {
+            get onDidChangeNativeLayout() {
                 return this.emitter.event;
             }
             getNativeLayout(): Promise<NativeKeyboardLayout> {
